test: Rodando migrações entre cada teste end to end

+++ src/nestjs/src/@share/testing/helpers.ts
@@ -1,4 +1,6 @@
+import { migrator } from '@fc/micro-videos/@seedwork/infra';
 import { INestApplication } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { getConnectionToken } from '@nestjs/sequelize';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AppModule } from '../../app.module';
@@ -8,14 +10,31 @@ export function startApp({
   beforeInit,
 }: { beforeInit?: (app: INestApplication) => void } = {}) {
   let _app: INestApplication;
+  let canRunMigrations: boolean;
 
   beforeEach(async () => {
     const moduleBuilder: TestingModule = await Test.createTestingModule({
       imports: [AppModule],
     }).compile();
 
+    canRunMigrations = !moduleBuilder
+      .get(ConfigService)
+      .get('DB_AUTO_LOAD_MODELS');
+
     const sequelize = moduleBuilder.get(getConnectionToken());
-    await sequelize.sync({ force: true });
+
+    try {
+      if (canRunMigrations) {
+        const umzug = migrator(sequelize, { logger: undefined });
+        await umzug.down({ to: 0 as any });
+        await umzug.up();
+      } else {
+        await sequelize.sync({ force: true });
+      }
+    } catch (e) {
+      console.log(e);
+      throw e;
+    }
 
     _app = moduleBuilder.createNestApplication();
     applyGlobalConfig(_app);
